# Conflicts:
#	src/main/java/com/stationery/project/interceptor/WriterCheckInterceptor.java

+++ src/main/webapp/WEB-INF/views/board/detail.jsp
@@ -6,6 +6,13 @@
 <head>
 <meta charset="UTF-8">
 <title>Insert title here</title>
+<style type="text/css">
+	img {
+	width : 200px;
+	height : 200px;
+	}
+
+</style>
 </head>
 <body>
 	<h1>${board} Detail Page</h1>
@@ -18,14 +25,19 @@
 		첨부사진
 		<c:forEach items="${dto.fileDTOs}" var="f"> <!-- 파일 다운로드 -->
 			<h3>${f.oriName}</h3>
+		<!-- 	<img alt="" src="../resources/upload/product/fileName(UUID)">  -->
+			<img alt="" src="../resources/upload/notices/${f.fileName}">
 		</c:forEach>
 	</c:if>
+	
+	<hr>
+	
 	<a href="./list">List</a>
-	<%-- <c:if test="${member.id eq dto.id}">
-		<a href="./delete?num=${dto.num}">Delete</a>
+	<!-- 작성자만 수정과 삭제가 가능하게끔  -->
+	<c:if test="${auth.id eq dto.id}">
+		<!-- <button type="button" onclick="button_del();">삭제하기</button> -->
+		<a href="./delete?num=${dto.num}" onclick="return confirm('정말 삭제하시겠습니까?');">Delete</a>
 		<a href="./update?num=${dto.num}">update</a>
-	</c:if> --%>
-	<a href="./delete?num=${dto.num}">Delete</a>
-	<a href="./update?num=${dto.num}">update</a>
+	</c:if>
 </body>
 </html>
